swtich to jump/crouch controls, optimize calls
