+++ src/QuantumElevators.csproj
@@ -65,12 +65,8 @@
     <Compile Include="ConsoleCmdQuantumElevators.cs" />
     <Compile Include="CoreLogic.cs" />
     <Compile Include="EntityAlive_Patches.cs" />
-    <Compile Include="disabled\EntityBuffs_Patches.cs" />
-    <Compile Include="disabled\EntityPlayerLocal_Patches.cs" />
-    <Compile Include="disabled\Entity_Patches.cs" />
     <Compile Include="ModApi.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TransportationServices.cs" />
     <Compile Include="ModLog.cs" />
     <Compile Include="QuantumCache.cs" />
   </ItemGroup>
